requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
pymongo>=4.8.0
motor>=3.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
openai>=1.35.0
python-dotenv>=1.0.1
requests>=2.32.0
